polish: move lily api to lens, reduce lotus api consumed
